Moves accordion to correct directory
